Improved Marching Cubes
Added a dictionary to meshing function, vertices which are already in the mesh are not re-added instead their index is added.
The mesh now has normals calculated.
Updated the marching cubes function to better utilize some c++ features.

+++ PA11/include/chunk.h
@@ -15,7 +15,7 @@ struct Chunk : public Object {
         };
 
         Type id;
-        double isoLevel;
+        float isoLevel;
     };
 
     // TODO: Chunk width
